+++ CHANGELOG.md
@@ -33,6 +33,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - automatic bonus damage dice modifier for skillful weapons
 - tag display in item sheet header
 - tag input field to edit free-form tags
+- RuleElements now show their selected documents in their sheets
 
 ### Changed